refactor: MAT-269 선수 할당 관련 어댑터를 실제 컴포넌트 옆에 두어 캡슐화 개선
